렌더러 방향성 전환
CPU(Legacy) 렌더러, D3D11(PBR) 렌더러, Vulkan(PBR) 렌더러, 레이트레이싱 렌더러를 구성하려던 방향성에서
CPU(PBR) 렌더링을 구현하는 방향성으로 전환되었음
D3D11 렌더러는 구현하지 않을 예정
Vulkan 렌더러는 다른 리포지토리에서 구현할 계획

+++ SimpleRenderer/SimpleRenderer/SimpleRenderer/CpuShader.h
@@ -53,7 +53,7 @@ struct PsInput
 // GPU에서 내부적으로 사용하는 메모리라고 생각합시다.
 
 ShaderConstants g_constants; // 쉐이더 상수
-std::vector<size_t> g_indexBuffer;
+std::vector<uint32_t> g_indexBuffer;
 std::vector<DirectX::SimpleMath::Vector4> g_displayBuffer;
 std::vector<float> g_depthBuffer;
 std::vector<DirectX::SimpleMath::Vector3> g_vertexBuffer;
